refactor task router by creating a controller for it
